+++ src/main/java/com/github/mjkuranda/spaceadventure2/SpaceAdventure2.java
@@ -1,13 +1,11 @@
 package com.github.mjkuranda.spaceadventure2;
 
-import com.github.mjkuranda.spaceadventure2.resources.GameAnimation;
-import com.github.mjkuranda.spaceadventure2.resources.GameFont;
-import com.github.mjkuranda.spaceadventure2.resources.GameImage;
 import com.github.mjkuranda.spaceadventure2.states.GameState;
 import com.github.mjkuranda.spaceadventure2.states.IntroState;
 import com.github.mjkuranda.spaceadventure2.states.LoadingState;
 import com.github.mjkuranda.spaceadventure2.states.highscore.HighScoreState;
 import com.github.mjkuranda.spaceadventure2.states.menus.GameOverMenuState;
+import com.github.mjkuranda.spaceadventure2.states.menus.HowToPlayMenuState;
 import com.github.mjkuranda.spaceadventure2.states.menus.MainMenuState;
 import com.github.mjkuranda.spaceadventure2.states.menus.NewGameMenuState;
 import org.newdawn.slick.*;
@@ -20,7 +18,7 @@ import java.util.logging.Logger;
 public class SpaceAdventure2 extends StateBasedGame {
 
     public static final String GAME_TITLE   = "Space Adventure II";
-    public static final String GAME_VERSION = "v0.21.0";
+    public static final String GAME_VERSION = "v0.22.0";
     public static final String GAME_CREATOR = "Marek Kuranda";
 
     public SpaceAdventure2() {
@@ -35,6 +33,7 @@ public class SpaceAdventure2 extends StateBasedGame {
         addState(new GameState());
         addState(new GameOverMenuState(this));
         addState(new HighScoreState(this));
+        addState(new HowToPlayMenuState(this));
         addState(new IntroState());
     }