+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.22.0] - 2023-08-14
+### Added
+- [@mjkuranda](https://github.com/mjkuranda): Fading out and in text label in main menu.
+- [@mjkuranda](https://github.com/mjkuranda): Fading out and in sign next to input string menu option.
+- [@mjkuranda](https://github.com/mjkuranda): Space screens animation from `Space Adventure 2019`.
+
+### Fixed
+- [@mjkuranda](https://github.com/mjkuranda): Loading resource names in `LoadingState`.
+
 ## [0.21.0] - 2023-08-12
 ### Added
 - [@mjkuranda](https://github.com/mjkuranda): `GameBar` component.